move term Obj classes to seperate file
